backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.0
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
python-multipart>=0.0.9
google-api-python-client>=2.116.0
google-auth>=2.27.0
httpx>=0.26.0
email-validator>=2.1.1
